docs: more SteamUIStore progress
